# Modules/Auth/Http/Requests/ResetPasswordRequest.php
<?php

namespace Modules\Auth\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;

class ResetPasswordRequest extends FormRequest
{
    public function authorize(): bool
    {
        return true;
    }

    public function rules(): array
    {
        return [
            'email'    => 'required|email|max:255',
            'token'    => 'required|string',
            'password' => 'required|string|min:8|confirmed',
        ];
    }
}